Improve decimal accuracy for rotation on properties panel
Fixes 90 being displayed as 89.999999.

+++ src/implui/contextrotation.cc
@@ -106,13 +106,13 @@ void ContextRotation::modelChanged( int changeBits )
 
       QString str;
 
-      str.sprintf( "%f", rad[0] );
+      str.sprintf( "%f", (float)rad[0] );
       m_xValue->setText( str );
 
-      str.sprintf( "%f", rad[1] );
+      str.sprintf( "%f", (float)rad[1] );
       m_yValue->setText( str );
 
-      str.sprintf( "%f", rad[2] );
+      str.sprintf( "%f", (float)rad[2] );
       m_zValue->setText( str );
 
       this->setEnabled( searching ? false : true );
